refactor frontend code
