+++ frontend/src/components/shared/Loaders/LoaderHelix.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useEffect } from 'react'
 
 export default function LoaderHelix() {
